Implemented the Xinco Setting admin feature

+++ trunk/XincoWebApp/src/java/com/bluecubs/xinco/core/server/XincoSettingServer.java
@@ -5,7 +5,6 @@ import com.bluecubs.xinco.core.server.persistence.controller.XincoSettingJpaCont
 import com.bluecubs.xinco.core.server.persistence.controller.exceptions.NonexistentEntityException;
 import com.bluecubs.xinco.core.server.service.XincoCoreUser;
 import com.bluecubs.xinco.core.server.service.XincoSetting;
-import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -32,7 +31,7 @@ public class XincoSettingServer extends XincoSetting {
             if (setting.getIntValue() != null) {
                 setIntValue(setting.getIntValue());
             }
-            setLongValue(setting.getLongValue().longValue());
+            setLongValue(setting.getLongValue());
             if (setting.getStringValue() != null) {
                 setStringValue(setting.getStringValue());
             }
@@ -63,7 +62,7 @@ public class XincoSettingServer extends XincoSetting {
                 setting = controller.findXincoSetting(getId());
                 setting.setBoolValue(isBoolValue());
                 setting.setIntValue(getIntValue());
-                setting.setLongValue(BigInteger.valueOf(getLongValue()));
+                setting.setLongValue(getLongValue());
                 setting.setStringValue(getStringValue());
                 setting.setDescription(getDescription());
                 controller.edit(setting);
@@ -71,7 +70,7 @@ public class XincoSettingServer extends XincoSetting {
                 setting = new com.bluecubs.xinco.core.server.persistence.XincoSetting();
                 setting.setBoolValue(isBoolValue());
                 setting.setIntValue(getIntValue());
-                setting.setLongValue(BigInteger.valueOf(getLongValue()));
+                setting.setLongValue(getLongValue());
                 setting.setStringValue(getStringValue());
                 setting.setDescription(getDescription());
                 controller.create(setting);
